backend/requirements.txt
asgiref==3.8.1
cryptography
Django==3.2.4
djangorestframework==3.15.1
djangorestframework-simplejwt
gunicorn
psycopg2-binary
pytz==2024.2
sqlparse==0.5.2
typing_extensions==4.12.2
